Finalizing the basic gameplay
Added the communication between soldiers, zombies, and players to correctly work the way that the game is designed. Soldier sees a zombie, asks for an item. The player then can deliver an item, if it is correct the soldier shoots and kills the zombie but if it is incorrect the soldier is stunned. Art assets on the box are still missing and the speech bubbles need to be modified to be clearer.